新增搜索功能（elasticsearch）

+++ youxiang/youxiang-common/src/main/java/com/youxiang/common/pojo/PageResult.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class PageResult<T> {
     private Long total;
-    private Long totalPage;
+    private Integer totalPage;
     private List<T> items;
 
     public PageResult(){
@@ -15,7 +15,7 @@ public class PageResult<T> {
         this.items=items;
     }
 
-    public PageResult(Long total,Long totalPage,List<T> items){
+    public PageResult(Long total,Integer totalPage,List<T> items){
         this.total=total;
         this.totalPage=totalPage;
         this.items=items;
@@ -29,11 +29,11 @@ public class PageResult<T> {
         this.total = total;
     }
 
-    public Long getTotalPage() {
+    public Integer getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Long totalPage) {
+    public void setTotalPage(Integer totalPage) {
         this.totalPage = totalPage;
     }
